LLM test first pass
