+++ packages/llm-completion/package.json
@@ -13,7 +13,7 @@
   },
   "scripts": {
     "build": "tsc -p tsconfig.json",
-    "test": "vitest"
+    "test": "vitest --run"
   },
   "engines": {
     "node": ">=22"
